% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagwood.R
\name{dagwood}
\alias{dagwood}
\title{DAGs with omitted objects displayed}
\usage{
dagwood(
  formula.DAG,
  exposure = NA,
  outcome = NA,
  formula.KUERs = NA,
  instrument = NA
)
}
\arguments{
\item{formula.DAG}{A string formula describing the DAG, in the style of DAGitty}

\item{exposure}{A string identifying the exposure of interest (must match the formula above)}

\item{outcome}{A string identifying the outcome of interest (must match the formula above)}

\item{formula.KUERs}{Not yet implemented}

\item{instrument}{Not yet implemented}
}
\description{
This function takes a root DAG and generates a DAGWOOD object from it. Details are available
from a preprint, here: https://arxiv.org/abs/2004.04251

Current limitations to this package:
Does not currently support instrumental variables (in progress).
Does not currently support non-minimal adjustment sets from root DAGs
Does not currently match KUER branch DAG matching with the UUER branch DAGs

Future features: improved graphical outputs, additional branch DAG types, possibly a GUI
Improve DAG import features and data checking
}
\examples{
formula.DAG <-"Chocolate -> Alzheimers
Chocolate <- Education -> Alzheimers
Chocolate -> CV
CV -> Alzheimers"
exposure <- "Chocolate"
outcome <- "Alzheimers"
branch.DAGs <- dagwood(formula.DAG,exposure,outcome)$DAGs.branch
}
\keyword{DAG,}
\keyword{causal}
\keyword{inference}
